third_party/WebKit/LayoutTests/external/wpt/content-security-policy/generic/policy-inherited-correctly-by-plznavigate-expected.txt

This is a testharness.js-based test.
PASS iframe still inherits correct CSP
FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
Harness: the test ran to completion.
